refactor: move SnapcraftBuildPlanner to projects model module
