+++ snapcraft/models/project.py
@@ -15,9 +15,9 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 """Project file definition and helpers."""
+from __future__ import annotations
 
 # pylint: disable=too-many-lines
-
 import copy
 import re
 from typing import (
@@ -28,6 +28,7 @@ from typing import (
     Literal,
     Mapping,
     Optional,
+    Sequence,
     Tuple,
     Union,
     cast,
@@ -43,7 +44,7 @@ from craft_providers import bases
 from pydantic import PrivateAttr, constr
 
 from snapcraft import utils
-from snapcraft.const import SnapArch
+from snapcraft.const import SUPPORTED_ARCHS, SnapArch
 from snapcraft.elf.elf_utils import get_arch_triplet
 from snapcraft.errors import ProjectValidationError
 from snapcraft.providers import SNAPCRAFT_BASE_TO_PROVIDER_BASE
@@ -1058,3 +1059,106 @@ def _format_global_keyword_warning(keyword: str, empty_entries: List[str]) -> st
         "stanza which is intended for configuration only."
         "\n(Reference: https://snapcraft.io/docs/snapcraft-interfaces)"
     )
+
+
+class SnapcraftBuildPlanner(models.BuildPlanner):
+    """A project model that creates build plans."""
+
+    base: str | None
+    build_base: str | None = None
+    name: str
+    platforms: dict[str, Any] | None = None
+    project_type: str | None = pydantic.Field(default=None, alias="type")
+
+    @pydantic.validator("platforms")
+    @classmethod
+    def _validate_all_platforms(cls, platforms: dict[str, Any]) -> dict[str, Any]:
+        """Validate and convert platform data to a dict of Platforms."""
+        for platform_label in platforms:
+            platform_data: dict[str, Any] = (
+                platforms[platform_label] if platforms[platform_label] else {}
+            )
+            error_prefix = f"Error for platform entry '{platform_label}'"
+
+            # Make sure the provided platform_set is valid
+            try:
+                platform = Platform(**platform_data)
+            except CraftValidationError as err:
+                raise CraftValidationError(f"{error_prefix}: {str(err)}") from None
+
+            # build_on and build_for are validated
+            # let's also validate the platform label
+            if platform.build_on:
+                build_on_one_of: Sequence[SnapArch | str] = platform.build_on
+            else:
+                build_on_one_of = [platform_label]
+
+            # If the label maps to a valid architecture and
+            # `build-for` is present, then both need to have the same value,
+            # otherwise the project is invalid.
+            if platform.build_for:
+                build_target = platform.build_for[0]
+                if platform_label in SUPPORTED_ARCHS and platform_label != build_target:
+                    raise CraftValidationError(
+                        str(
+                            f"{error_prefix}: if 'build_for' is provided and the "
+                            "platform entry label corresponds to a valid architecture, then "
+                            f"both values must match. {platform_label} != {build_target}"
+                        )
+                    )
+            # if no build-for is present, then the platform label needs to be a valid architecture
+            elif platform_label not in SUPPORTED_ARCHS:
+                raise CraftValidationError(
+                    str(
+                        f"{error_prefix}: platform entry label must correspond to a "
+                        "valid architecture if 'build-for' is not provided."
+                    )
+                )
+
+            # Both build and target architectures must be supported
+            if not any(b_o in SUPPORTED_ARCHS for b_o in build_on_one_of):
+                raise CraftValidationError(
+                    str(
+                        f"{error_prefix}: trying to build snap in one of "
+                        f"{build_on_one_of}, but none of these build architectures are supported. "
+                        f"Supported architectures: {SUPPORTED_ARCHS}"
+                    )
+                )
+
+            platforms[platform_label] = platform
+
+        return platforms
+
+    def get_build_plan(self) -> List[BuildInfo]:
+        """Get the build plan for this project."""
+        build_infos: list[BuildInfo] = []
+        effective_base = SNAPCRAFT_BASE_TO_PROVIDER_BASE[
+            str(
+                get_effective_base(
+                    base=self.base,
+                    build_base=self.build_base,
+                    project_type=self.project_type,
+                    name=self.name,
+                )
+            )
+        ].value
+
+        base = bases.BaseName("ubuntu", effective_base)
+
+        # set default value
+        if self.platforms is None:
+            self.platforms = {get_host_architecture(): None}
+
+        for platform_entry, platform in self.platforms.items():
+            for build_for in platform.build_for or [platform_entry]:
+                for build_on in platform.build_on or [platform_entry]:
+                    build_infos.append(
+                        BuildInfo(
+                            platform=platform_entry,
+                            build_on=build_on,
+                            build_for=build_for,
+                            base=base,
+                        )
+                    )
+
+        return build_infos
